Adds support for PAMI Star robot group
Adds the necessary components and configuration to support the "PAMI Star" robot group, similar to the existing triangle, square, and round groups.

This includes defining the `InitStep` for star calibration, creating a `RobotGroup` and `RobotGroupService` for the star group, and integrating it into the `NerellOrdonanceur` to manage its lifecycle and communication.
Also, corrects a typo in the `InitStep` enum.

+++ robot-system-lib-parent/robot-system-lib-core/src/main/java/org/arig/robot/scheduler/TasksScheduler.java
@@ -103,7 +103,7 @@ public class TasksScheduler implements InitializingBean {
                 long timeStartI2C = System.nanoTime();
                 long ellapsedI2C = timeStartI2C - lastTimeI2C;
 
-                if (ellapsedI2C >= robotConfig.i2cReadTimeMs() * MS_TO_NS) {
+                if (ellapsedI2C >= (long) robotConfig.i2cReadTimeMs() * MS_TO_NS) {
                     lastTimeI2C = timeStartI2C;
 
                     ioService.refreshAllIO();
